requirements.txt
streamlit
pandas
numpy
scikit-learn
pdfplumber
tabula-py
pytesseract
openpyxl
matplotlib
seaborn
joblib
sentence-transformers
plotly
